+++ nox/src/node.rs
@@ -40,7 +40,7 @@ use libp2p_connection_limits::ConnectionLimits;
 use libp2p_metrics::{Metrics, Recorder};
 use particle_builtins::{Builtins, CustomService, NodeInfo};
 use particle_execution::ParticleFunctionStatic;
-use particle_protocol::Particle;
+use particle_protocol::ExtendedParticle;
 use peer_metrics::{
     ConnectionPoolMetrics, ConnectivityMetrics, ParticleExecutorMetrics, ServicesMetrics,
     ServicesMetricsBackend, SpellMetrics, VmPoolMetrics,
@@ -67,7 +67,7 @@ use crate::metrics::TokioCollector;
 
 // TODO: documentation
 pub struct Node<RT: AquaRuntime> {
-    particle_stream: mpsc::Receiver<Particle>,
+    particle_stream: mpsc::Receiver<ExtendedParticle>,
     effects_stream: mpsc::Receiver<Result<RoutingEffects, AquamarineApiError>>,
     pub swarm: Swarm<FluenceNetworkBehaviour>,
 
@@ -192,6 +192,7 @@ impl<RT: AquaRuntime> Node<RT> {
             transport,
             config.external_addresses(),
             health_registry.as_mut(),
+            metrics_registry.as_mut(),
         )?;
 
         let (services_metrics_backend, services_metrics) =
@@ -353,23 +354,32 @@ impl<RT: AquaRuntime> Node<RT> {
         transport: Boxed<(PeerId, StreamMuxerBox)>,
         external_addresses: Vec<Multiaddr>,
         health_registry: Option<&mut HealthCheckRegistry>,
+        metrics_registry: Option<&mut Registry>,
     ) -> eyre::Result<(
         Swarm<FluenceNetworkBehaviour>,
         Connectivity,
-        mpsc::Receiver<Particle>,
+        mpsc::Receiver<ExtendedParticle>,
     )> {
         let connection_idle_timeout = network_config.connection_idle_timeout;
 
         let (behaviour, connectivity, particle_stream) =
             FluenceNetworkBehaviour::new(network_config, health_registry);
 
-        let mut swarm = SwarmBuilder::with_existing_identity(key_pair)
-            .with_tokio()
-            .with_other_transport(|_| transport)?
-            .with_behaviour(|_| behaviour)?
-            .with_swarm_config(|cfg| cfg.with_idle_connection_timeout(connection_idle_timeout))
-            .build();
-
+        let mut swarm = match metrics_registry {
+            None => SwarmBuilder::with_existing_identity(key_pair)
+                .with_tokio()
+                .with_other_transport(|_| transport)?
+                .with_behaviour(|_| behaviour)?
+                .with_swarm_config(|cfg| cfg.with_idle_connection_timeout(connection_idle_timeout))
+                .build(),
+            Some(registry) => SwarmBuilder::with_existing_identity(key_pair)
+                .with_tokio()
+                .with_other_transport(|_| transport)?
+                .with_bandwidth_metrics(registry)
+                .with_behaviour(|_| behaviour)?
+                .with_swarm_config(|cfg| cfg.with_idle_connection_timeout(connection_idle_timeout))
+                .build(),
+        };
         // Add external addresses to Swarm
         external_addresses.iter().cloned().for_each(|addr| {
             Swarm::add_external_address(&mut swarm, addr);
@@ -404,7 +414,7 @@ pub struct StartedNode {
 impl<RT: AquaRuntime> Node<RT> {
     #[allow(clippy::too_many_arguments)]
     pub fn with(
-        particle_stream: mpsc::Receiver<Particle>,
+        particle_stream: mpsc::Receiver<ExtendedParticle>,
         effects_stream: mpsc::Receiver<Result<RoutingEffects, AquamarineApiError>>,
         swarm: Swarm<FluenceNetworkBehaviour>,
         connectivity: Connectivity,